requirements.txt
aiohttp
matplotlib
numpy
pandas
pytz
requests
pynacl
python-dotenv
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
discord.py>=2.3.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
bandit>=1.7.0
safety>=2.3.0